requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
PyJWT==2.7.0
qrcode==7.4.2
Pillow==9.5.0
email_validator==2.0.0
python-dotenv==1.0.0
Flask-Session==0.5.0
Flask-Admin==1.6.1
pyotp==2.8.0
pytest==7.3.1
